Rename helper.py, separate utils from Card class
